Agregado de cambio de roles y rol premium

+++ src/controllers/product.controller.js
@@ -1,5 +1,6 @@
 import productService from "../dao/services/product.service.js";
 import cartService from "../dao/services/cart.service.js";
+import userService from "../dao/services/user.service.js";
 
 const productController = {
     getProducts: async (req, res) => {
@@ -88,11 +89,21 @@ const productController = {
     updateProduct: async (req, res) => {
         const productId = req.params.pid;
         const productUpdateData = req.body;
+        const userId = req.session.userId;
+        const userRole = req.session.userRole;
     
         try {
-            const updatedProduct = await productService.updateProduct(productId, req, productUpdateData);
-    
-            return res.json({ message: "Producto actualizado!", product: updatedProduct });
+            const product = await productService.getProductDetail(productId);
+            const user = await userService.getUserById(userId)
+
+            if (userRole == 'admin' || (userRole == 'premium' && user == product.owner)) {
+                const updatedProduct = await productService.updateProduct(productId, req, productUpdateData);
+
+                return res.json({ message: "Producto actualizado!", product: updatedProduct });
+            }
+            else{
+                return res.status(403).json({ message: 'No tienes permiso para realizar esta acción' });
+            }
         } catch (err) {
             console.error('Error:', err);
             return res.status(500).json({ error: "Error en la base de datos", details: err.message });
@@ -119,11 +130,21 @@ const productController = {
 
     deleteProduct: async (req, res) => {
         const productId = req.params.pid;
+        const userId = req.session.userId;
+        const userRole = req.session.userRole;
 
         try {
-            await productService.deleteProduct(productId);
+            const product = await productService.getProductDetail(productId);
+            const user = await userService.getUserById(userId)
 
-            return res.json({ message: "Producto eliminado!" });
+            if (userRole == 'admin' || (userRole == 'premium' && user == product.owner)) {
+                await productService.deleteProduct(productId);
+
+                return res.json({ message: "Producto eliminado!" });
+            }
+            else{
+                return res.status(403).json({ message: 'No tienes permiso para realizar esta acción' });
+            }
         } catch (err) {
             console.error('Error:', err);
             return res.status(500).json({ error: "Error en la base de datos", details: err.message });
